refactor: reorganize module structure and improve AST data handling
- Split code into parser/, renderer/, ast/ module structure
- Fix data_num() to use u64 instead of u8
- Fix repeat() calls with proper usize conversion
- Remove unused imports and cfg warnings
- Improve error handling with thiserror derive

+++ src/parser/mdx/mod.rs
@@ -259,7 +259,7 @@ fn convert_node(node: &MdastNode) -> Vec<Node> {
             let content = convert_inlines(&heading.children);
             vec![Node::new("heading")
                 .with_children(content)
-                .data_num("depth", heading.depth as u8)]
+                .data_num("depth", heading.depth as u64)]
         }
         MdastNode::Paragraph(para) => {
             let content = convert_inlines(&para.children);
